Completion of the  main fuctions: read, write, search
