+++ custom_classes/exception_classes.py
