+++ service_functions/__pycache__/file_operations.cpython-310.pyc
