+++ service_functions/__pycache__/get_info.cpython-310.pyc
